@@ -1,46 +1,35 @@
 {
     "name": "laravel/laravel",
     "type": "project",
-    "description": "The Laravel Framework.",
-    "keywords": [
-        "framework",
-        "laravel"
-    ],
+    "description": "The skeleton application for the Laravel framework.",
+    "keywords": ["laravel", "framework"],
     "license": "MIT",
     "require": {
         "php": "^8.1",
-        "artesaos/seotools": "^1.1",
-        "astrotomic/laravel-translatable": "^11.12",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^10.0",
-        "laravel/octane": "^1.5",
+        "inertiajs/inertia-laravel": "^0.6.3",
+        "laravel/framework": "^10.10",
+        "laravel/octane": "^2.2",
         "laravel/sanctum": "^3.2",
-        "laravel/serializable-closure": "^1.3",
         "laravel/tinker": "^2.8",
-        "livewire/livewire": "^3.0.1",
-        "mhmiton/laravel-modules-livewire": "^1.6",
-        "nwidart/laravel-modules": "^10.0",
-        "realrashid/sweet-alert": "^6.0"
+        "tightenco/ziggy": "^1.0"
     },
     "require-dev": {
         "fakerphp/faker": "^1.9.1",
+        "laravel/breeze": "^1.28",
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^7.0",
-        "phpunit/phpunit": "^10.0",
+        "phpunit/phpunit": "^10.1",
         "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
             "App\\": "app/",
-            "Modules\\": "Modules/",
             "Database\\Factories\\": "database/factories/",
             "Database\\Seeders\\": "database/seeders/"
-        },
-        "files": [
-            "app/Helpers.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
