@@ -2,12 +2,8 @@
 
 namespace App\Providers;
 
-use App\Models\Permission;
-use App\Models\User;
+// use Illuminate\Support\Facades\Gate;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Facades\Gate;
-use Illuminate\Support\Facades\Schema;
 
 class AuthServiceProvider extends ServiceProvider
 {
@@ -17,7 +13,7 @@ class AuthServiceProvider extends ServiceProvider
      * @var array<class-string, class-string>
      */
     protected $policies = [
-        // 'App\Models\Model' => 'App\Policies\ModelPolicy',
+        //
     ];
 
     /**
@@ -25,39 +21,6 @@ class AuthServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        $this->registerPolicies();
-
-        $this->app->scoped('userPermissions', function () {
-            $user = $this->getUser();
-
-            if (is_null($user)) return [];
-
-            // if user has no role , return empty array
-            return $user->role()
-                ->with('permissions')
-                ->first()
-                ?->permissions
-                ->pluck('tag')
-                ->toArray() ?? [];
-        });
-
-        if (Schema::hasTable('permissions')) {
-            $permissions = Cache::remember(
-                'allPermissions',
-                14400,
-                fn () => Permission::select(['id', 'tag'])->get()->pluck('tag')->toArray()
-            );
-
-            foreach ($permissions as $permission) {
-                Gate::define($permission, function () use ($permission) {
-                    return in_array($permission, app('userPermissions'));
-                });
-            }
-        }
-    }
-
-    protected function getUser(): ?User
-    {
-        return auth('sanctum')->user() ?? auth('web')->user() ?? null;
+        //
     }
 }
