@@ -13,13 +13,15 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('email')->nullable()->unique();
+            $table->unsignedBigInteger('role_id')->nullable();
+            $table->foreign('role_id')->references('id')->on('roles')->onDelete('set null');
+            $table->string('first_name');
+            $table->string('last_name')->nullable();
             $table->string('number')->nullable()->unique();
-            $table->string('full_name');
+            $table->string('email')->nullable()->unique();
+            $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('role_id')->default(0);
             $table->rememberToken();
-            $table->timestamp('verified_at')->nullable();
             $table->timestamps();
         });
     }
