@@ -1,8 +1,6 @@
-module.exports = {
-  plugins: [
-    require('postcss-import'),
-    require('@tailwindcss/nesting'),
-    require('tailwindcss'),
-    require('autoprefixer'),
-  ],
-}
+export default {
+    plugins: {
+        tailwindcss: {},
+        autoprefixer: {},
+    },
+};
