@@ -1,27 +1,22 @@
+import defaultTheme from 'tailwindcss/defaultTheme';
+import forms from '@tailwindcss/forms';
+
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./Modules/**/Resources/views/**/*.blade.php",
-    "./Modules/**/Resources/views/*.blade.php",
-    "./resources/**/*.blade.php",
-    "./resources/**/*.js",
-    "./resources/**/*.vue",
-  ],
-  theme: {
-    extend: {
-      colors: {
-        theme: {
-          "secondary": "#f8f9ff",
-          "primary": "#5c68ff",
-          "darked-primary": "#4d56e0",
-          "gray": "#fafafa",
+export default {
+    content: [
+        './vendor/laravel/framework/src/Illuminate/Pagination/resources/views/*.blade.php',
+        './storage/framework/views/*.php',
+        './resources/views/**/*.blade.php',
+        './resources/js/**/*.jsx',
+    ],
+
+    theme: {
+        extend: {
+            fontFamily: {
+                sans: ['Figtree', ...defaultTheme.fontFamily.sans],
+            },
         },
-      },
     },
-  },
-  variants: {
-    extend: {},
-  },
-  plugins: [],
-  corePlugins: {},
+
+    plugins: [forms],
 };
