@@ -1,16 +1,13 @@
 import { defineConfig } from 'vite';
 import laravel from 'laravel-vite-plugin';
+import react from '@vitejs/plugin-react';
 
 export default defineConfig({
     plugins: [
         laravel({
-            input: [
-                'resources/css/app.css', 
-                'resources/views/panel/layouts/ZedAdmin/src/css/app.css',
-                'resources/views/panel/layouts/ZedAdmin/src/js/dashboard.js',
-                'resources/js/app.js'
-            ],
+            input: 'resources/js/app.jsx',
             refresh: true,
         }),
+        react(),
     ],
 });
